Correctly initialize nodes array
We got _very_ lucky with how we initialized the nodes array.  That sort
of structure doesn't work for numpy in almost every other case.  In
particular, it messes everything up when copying.

This wasn't yet causing us problems, but it did cause me problems when I
tried to do exactly the same thing with an array that didn't start as a
list of [0,0] elements.